第一次完赛（无圆环版）

+++ project/code/control.h
@@ -6,8 +6,8 @@
 extern uint8 Motor_Protection_flag;
 
 
-void Motor_Left_PWM(int8_t duty);
-void Motor_Right_PWM(int8_t duty);
+void Motor_Left_PWM(int16_t duty);
+void Motor_Right_PWM(int16_t duty);
 void Motor_Protection(void);
